requirements.txt
pyyaml=21.3.1